+++ v2/pkg/ast/ast_value.go
@@ -174,14 +174,27 @@ func (d *Document) writeJSONValue(buf *bytes.Buffer, value Value) error {
 		buf.WriteByte(literal.RBRACK_BYTE)
 	case ValueKindObject:
 		buf.WriteByte(literal.LBRACE_BYTE)
+
+		hasRenderedFields := false
 		for ii, ref := range d.ObjectValues[value.Ref].Refs {
-			if ii > 0 {
+			objFieldValue := d.ObjectFieldValue(ref)
+			if objFieldValue.Kind == ValueKindVariable {
+				variableName := d.Input.ByteSliceString(d.VariableValues[objFieldValue.Ref].Name)
+				_, dataType, _, _ := jsonparser.Get(d.Input.Variables, variableName)
+				if dataType == jsonparser.NotExist {
+					continue
+				}
+			}
+
+			if ii > 0 && hasRenderedFields {
 				buf.WriteByte(literal.COMMA_BYTE)
 			}
+			hasRenderedFields = true
+
 			fieldNameBytes := d.ObjectFieldNameBytes(ref)
 			buf.Write(quotes.WrapBytes(fieldNameBytes))
 			buf.WriteByte(literal.COLON_BYTE)
-			if err := d.writeJSONValue(buf, d.ObjectFieldValue(ref)); err != nil {
+			if err := d.writeJSONValue(buf, objFieldValue); err != nil {
 				return err
 			}
 		}