+++ v2/pkg/engine/plan/configuration.go
@@ -12,6 +12,7 @@ type Configuration struct {
 	DataSources                []DataSource
 	Fields                     FieldConfigurations
 	Types                      TypeConfigurations
+	EntityInterfaceNames       []string
 	// DisableResolveFieldPositions should be set to true for testing purposes
 	// This setting removes position information from all fields
 	// In production, this should be set to false so that error messages are easier to understand