+++ v2/pkg/errorcodes/const.go
@@ -1,8 +1,12 @@
 package errorcodes
 
 const (
+	// Apollo Gateway Compatibility
 	BadUserInput            = "BAD_USER_INPUT"
 	GraphQLValidationFailed = "GRAPHQL_VALIDATION_FAILED"
 	InternalServerError     = "INTERNAL_SERVER_ERROR"
 	InvalidGraphql          = "INVALID_GRAPHQL"
+
+	// Apollo Router Compatibility
+	ValidationInvalidTypeVariable = "VALIDATION_INVALID_TYPE_VARIABLE"
 )